pyproject.toml
[project]
name = "spotify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client==2.166.0",
    "pytube==15.0.0",
    "yt-dlp==2025.3.31",
    "python-dotenv==1.1.0"
]
